+++ Cargo.toml
@@ -32,3 +32,4 @@ field_types = "1.1.0"
 function_name = "0.3.0"
 convert_case = "0.6.0"
 clap-markdown = "0.1.4"
+ambassador = "0.4.1"
