+++ src/beyond_identity/api/mod.rs
@@ -1,4 +1,5 @@
+pub mod commands;
+pub mod common;
 pub mod groups;
 pub mod identities;
 pub mod roles;
-pub mod common;
