+++ src/beyond_identity/mod.rs
@@ -1,6 +1,7 @@
 pub mod admin;
 pub mod api;
 pub mod api_token;
+pub mod commands;
 pub mod enrollment;
 pub mod external_sso;
 pub mod groups;
