+++ src/common/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
 pub mod error;
 pub mod http;
+pub mod command;
