+++ src/okta/mod.rs
@@ -3,3 +3,4 @@ pub mod identity_provider;
 pub mod registration_attribute;
 pub mod routing_rule;
 pub mod scim;
+pub mod commands;
