+++ src/onelogin/mod.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod fast_migrate;
